Fixed an int overflow bug in GetIndustryJobs.

+++ EveLib.EveOnline/Models/Character/IndustryJobs.cs
@@ -62,7 +62,7 @@ namespace eZet.EveLib.Modules.Models.Character {
             /// The station ID
             /// </summary>
             [XmlAttribute("stationID")]
-            public int StationId { get; set; }
+            public long StationId { get; set; }
 
             /// <summary>
             /// The activity ID
